align snippets layout with docs layout (under se)
