+++ docs/src/main/asciidoc/se/security/providers.adoc
@@ -78,7 +78,7 @@ This is not required when `redirect` is set to false.
 [source,java]
 .Adding support for OIDC redirects
 ----
-include::{sourcedir}/se/SecuritySnippets.java[tag=snippet_1, indent=0]
+include::{sourcedir}/se/security/ProvidersSnippets.java[tag=snippet_1, indent=0]
 ----
 
 include::{rootdir}/includes/security/providers/oidc.adoc[]
